internal/metrics: add metrics http handler
